renamed invoicesStore in github
Hopefully this fixes my netlify deploy. It seems that changing the name in VSC and commiting it didn't update the casing of the name in my repo. This should update it....

Netlify turns out to be case-sensitive when it comes to JS files. So InvoicesStore.js was breaking my deploy.